refactor(agents-ui): split AgentsApp.tsx into apps/agents/ components (B6a) (#558)
